Run 'npm' update
Updates client and server packages to highest minor version numbers.
This resolves a dependency conflict were some packages require different
versions of react and therefore makes it possible to build and run the
client with newer versions of nodejs. For the site to launch, it is
necessary to add the arg '--openssl-legacy-provider' to the
react-scripts start command. Upgrading the stack to use newer openssl
implementations may be a more involved task.