Add auth0 roles and make CreateEvent admin-only
Adds an auth0 action and instructions in the README for configuring
auth0 to use that action to give the application information about
whether a user has the admin role or not. Modifies the Navbar so that a
link to the CreateEvent component is only visible if the user is an
admin. Modifies the CreateEvent component so that if an unauthorized
user navigates to the `/create` route, an alternate component will be
rendered. Note that this commit doesn't do anything to secure access to
the backend API endpoints that CreateEvent uses.