Implement client API access token retrieval
The strategy for access token retrieval and use implemeneted here is to
get an access token once when the user logs in and the auth0 instance
starts up, and then add that token to an intercepter in a global axios
client via a useEffect React hook in the root App component. This way,
the we don't need to request a new token for every API call.

References:
- https://auth0.com/docs/quickstart/spa/react/02-calling-an-api
- https://stackoverflow.com/a/71558686

The CreateEvent component has been configured to use this global HTTP
client so that authorized users can create events using the `POST
/lessons` API endpoint.