Make fields on User model optional
In the current implementation, a user can sign up and get access to the
frontend application without providing any information other than an
email for Auth0 signup. I'm changing firstName and phoneNumber so they
aren't required, and adding a default value for userType. This way, a
User entity in the backend database can be created as soon as Auth0
registration is completed, with remaining information to be filled in
afterwards.