Add error handling to POST /users
The POST /users endpoint will now return a 400 error in case it is given
invalid data to create a user with, and a 500 status code in case of
other unforeseen errors. This commit also adds testing utilities for
connecting to a mock in memory MongoDB, and adds some unit tests
verifying the functionality of the POST /users endpoint.