Rework GET /users for consistency
Modifies GET /users so that it can be used to retrieve users using any
of the properties of the User model as a filter. Adds error
handling so that queries returning nothing yield a 404 error, and other
errors yield a 500 error. Also some unit tests for the GET /users
endpoint.