Add httpClient wrapping axios instance to server
The server will need to talk to the Auth0 management API. I've opted to
use a similar strategy to the one that's currently used in the client
for getting and using access tokens. There will need to be a mechanism
that the server can use to refresh the token periodically.