Add '/me' endpoint for normal user write access
This commit adds a new api endpoint at path /me that allows
authenticated users to register with the application database or modify
their profile information. It changes the Signup page to use this new
endpoint, and a new profile page thhat loads profile information from
the database and provides facilities for editing it. A new environment
variable was added to keep track of URL of the API server, as documented
in the readme. The POST, PUT, and DELETE verbs on the /users endpoint
will be reserved for users with the admin role in the future.