Configure /me endpoint to write to Auth0 db
This commit sets the application up to obtain access to the Auth0
management API, and use it in the case where a user changes their email
address through the /profile page on the client. Importantly, this fixes
the server's httpClient, which initially implemented incorrectly and did
not retrieve a valid token.