Integrate /users endpoint with relevant Auth0 APIs
The GET /users endpoint now makes two calls into the Auth0 api to
retrieve information about when users last logged in, and whether or not
they have the admin role. Integrated the new mock Auth0 server into the
tests and changed them to reflect new functionality. The README was
updated to include information about new Auth0 management API
permissions that need to be given to the application for the new
functionality to work. Also updated the /user routing configuration so
that only admin users can mutate user data.