Implement admin UI using MUI data-grid
Displays information from all relevant fields of the server database as
well as additional information from the Auth0 management API in a table.
The MUI data-grid comes with many built-in features such as sorting and
filtering. Where applicable, fields can be double-clicked to toggle edit
mode, but for now this doesn't propogate any changes anywhere. This is a
read-only view.