Implement DELETE /user/:id endpoint
This commit implements the DELETE /user/:id endpoint and integrates it
with the Auth0 management API. This is accomplished with a mongoose
schema hook that runs just prior to a deleteOne operation on any
instance of the User model, aborting if unsuccessful. I also made fixes
to the error handling in the route, and to how the program loads
environment variables depending on whether or not it is in a production
context.