Add mui x date-pickers and switch moment to dayjs
Adds material UI X datepickers to be used in the event creation form
instead of the older datepicker components that use the browser's native
input controls (TextField components with type="date"). This also seemed
like a good opportunity to switch from the moment datetime library,
which is in maintenance mode, to the more modern dayjs library. Dayjs
has mostly the same API as moment but with better performance and less
size. Also added the React testing library.